backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
black==22.3.0
click==8.1.3
fastapi==0.78.0
fuzzywuzzy==0.18.0
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
numpy==1.22.4
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.17.6
